Moved exception classes to domain
This allows the classes to be transported between the backend and the
web application using AkkaTransport.
