+++ geoide-document-cache/src/main/java/nl/idgis/geoide/documentcache/service/FileStore.java
@@ -13,7 +13,7 @@ import java.util.concurrent.CompletableFuture;
 import nl.idgis.geoide.commons.domain.MimeContentType;
 import nl.idgis.geoide.commons.domain.api.DocumentStore;
 import nl.idgis.geoide.commons.domain.document.Document;
-import nl.idgis.geoide.documentcache.DocumentCacheException;
+import nl.idgis.geoide.commons.domain.document.DocumentCacheException;
 import nl.idgis.geoide.util.Futures;
 import nl.idgis.geoide.util.streams.StreamProcessor;
