+++ geoide-document-cache/src/main/java/nl/idgis/geoide/documentcache/service/HttpDocumentStore.java
@@ -10,9 +10,9 @@ import nl.idgis.geoide.commons.domain.MimeContentType;
 import nl.idgis.geoide.commons.domain.api.DocumentCache;
 import nl.idgis.geoide.commons.domain.api.DocumentStore;
 import nl.idgis.geoide.commons.domain.document.Document;
+import nl.idgis.geoide.commons.domain.document.DocumentCacheException;
 import nl.idgis.geoide.commons.http.client.HttpClient;
 import nl.idgis.geoide.commons.http.client.HttpRequestBuilder;
-import nl.idgis.geoide.documentcache.DocumentCacheException;
 import nl.idgis.geoide.util.Futures;
 
 import org.reactivestreams.Publisher;
