+++ geoide-document-cache/src/test/java/nl/idgis/geoide/documentcache/service/TestDefaultDocumentCache.java
@@ -15,8 +15,8 @@ import java.util.concurrent.TimeUnit;
 import nl.idgis.geoide.commons.domain.MimeContentType;
 import nl.idgis.geoide.commons.domain.api.DocumentStore;
 import nl.idgis.geoide.commons.domain.document.Document;
-import nl.idgis.geoide.documentcache.DocumentCacheException;
-import nl.idgis.geoide.documentcache.DocumentCacheException.DocumentNotFoundException;
+import nl.idgis.geoide.commons.domain.document.DocumentCacheException;
+import nl.idgis.geoide.commons.domain.document.DocumentCacheException.DocumentNotFoundException;
 import nl.idgis.geoide.util.Futures;
 import nl.idgis.geoide.util.streams.AkkaStreamProcessor;
 import nl.idgis.geoide.util.streams.StreamProcessor;
