+++ geoide-print/src/main/java/nl/idgis/geoide/commons/print/service/DelegatingPrintService.java
@@ -15,9 +15,10 @@ import nl.idgis.geoide.commons.domain.MimeContentType;
 import nl.idgis.geoide.commons.domain.api.PrintService;
 import nl.idgis.geoide.commons.domain.document.Document;
 import nl.idgis.geoide.commons.domain.print.Capabilities;
+import nl.idgis.geoide.commons.domain.print.PrintException;
 import nl.idgis.geoide.commons.domain.print.PrintRequest;
 import nl.idgis.geoide.commons.domain.print.Capabilities.InputFormat;
-import nl.idgis.geoide.commons.print.service.PrintException.UnsupportedFormat;
+import nl.idgis.geoide.commons.domain.print.PrintException.UnsupportedFormat;
 import nl.idgis.geoide.util.Futures;
 
 /**
