+++ geoide-print/src/test/java/nl/idgis/geoide/commons/print/service/TestHtmlPrintService.java
@@ -18,8 +18,9 @@ import nl.idgis.geoide.commons.domain.MimeContentType;
 import nl.idgis.geoide.commons.domain.api.DocumentCache;
 import nl.idgis.geoide.commons.domain.document.Document;
 import nl.idgis.geoide.commons.domain.print.DocumentReference;
+import nl.idgis.geoide.commons.domain.print.PrintException;
 import nl.idgis.geoide.commons.domain.print.PrintRequest;
-import nl.idgis.geoide.commons.report.layout.less.LessCompilationException;
+import nl.idgis.geoide.commons.domain.report.LessCompilationException;
 import nl.idgis.geoide.documentcache.service.DefaultDocumentCache;
 import nl.idgis.geoide.util.Futures;
 import nl.idgis.geoide.util.streams.AkkaStreamProcessor;
